Fix for unit test runtime performance (#358)
* Fix for unit test speeds


* Fix spark tests


* Test fixes


* Change parameters


* Updates for tests


* Remove comments


* Update to CICD


* Add mkdocs test


---------
