+++ .github/workflows/sonarcloud_reusable.yml
@@ -48,7 +48,7 @@ jobs:
       matrix:
         os: [ubuntu-latest]
         python-version: ["3.11"]
-        pyspark: ["3.4.0"]
+        pyspark: ["3.4.1"]
         delta-spark: ["2.4.0"]
     runs-on: ${{ matrix.os }} 
     steps:
