+++ .github/workflows/test.yml
@@ -28,6 +28,13 @@ jobs:
           os: [ubuntu-latest]
           python-version: ["3.8", "3.9", "3.10", "3.11"]
           pyspark: ["3.3.0", "3.3.1", "3.3.2", "3.4.0"]
+          exclude:
+            - pyspark: "3.4.0"
+              python-version: "3.8"
+            - pyspark: "3.4.0"
+              python-version: "3.9"
+            - pyspark: "3.4.0"
+              python-version: "3.10"              
           include:
             - pyspark: "3.3.0"
               delta-spark: "2.2.0"
@@ -74,3 +81,50 @@ jobs:
             coverage run -m pytest --junitxml=xunit-reports/xunit-result-unitttests.xml tests
             coverage xml --omit "venv/**,maintenance/**,xunit-reports/**" -i -o coverage-reports-previous/coverage-unittests.xml
             echo Coverage `coverage report --omit "venv/**" | grep TOTAL | tr -s ' ' | cut -d" " -f4`
+
+  job_test_mkdocs:
+    defaults:
+      run:
+        shell: bash -l {0}
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        python-version: ["3.11"]
+        pyspark: ["3.4.1"]
+        delta-spark: ["2.4.0"]
+    runs-on: ${{ matrix.os }} 
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          repository: ${{ inputs.REPO_NAME }}
+          ref: ${{ inputs.HEAD_BRANCH }}
+          fetch-depth: 0
+
+      - name: Setup Python 
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      
+      - name: Install Boost
+        run: |
+          sudo apt update
+          sudo apt install -y libboost-all-dev
+
+      - name: Add conda to system path
+        run: |
+          # $CONDA is an environment variable pointing to the root of the miniconda directory
+          echo $CONDA/bin >> $GITHUB_PATH
+
+      - name: Install Conda environment with Micromamba
+        uses: mamba-org/setup-micromamba@main
+        with:
+          environment-file: environment.yml
+          create-args: >-
+            python=${{ matrix.python-version }}
+            pyspark=${{ matrix.pyspark }}
+            delta-spark=${{ matrix.delta-spark }}          
+          cache-environment: true
+
+      - name: Mkdocs Test
+        run: |
+          mkdocs build --strict
