+++ tests/sdk/python/rtdip_sdk/pipelines/_pipeline_utils/spark_configuration_constants.py
@@ -11,43 +11,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import sys
-
-from pytest_mock import MockerFixture
-sys.path.insert(0, '.')
-from typing import Iterator
-import pytest
 import os
 from pathlib import Path
 import shutil
 from dataclasses import dataclass
-from unittest.mock import patch
-
-from src.sdk.python.rtdip_sdk.pipelines.destinations import * # NOSONAR
-from src.sdk.python.rtdip_sdk.pipelines.sources import * # NOSONAR
-from src.sdk.python.rtdip_sdk.pipelines.utilities.spark.session import SparkSessionUtility
-
-SPARK_TESTING_CONFIGURATION = {
-    "spark.executor.cores": "2",
-    "spark.executor.instances": "2",
-    "spark.sql.shuffle.partitions": "2",
-    "spark.app.name": "test_app", 
-    "spark.master": "local[*]"
-}
-
-@pytest.fixture(scope="session")
-def spark_session():
-    spark = SparkSessionUtility(SPARK_TESTING_CONFIGURATION.copy()).execute()
-    path = spark.conf.get("spark.sql.warehouse.dir")
-    prefix = "file:"
-    if path.startswith(prefix):
-        path = path[len(prefix):]    
-    if os.path.isdir(path):
-        shutil.rmtree(path)    
-    yield spark
-    spark.stop()
-    if os.path.isdir(path):
-        shutil.rmtree(path)
 
 @dataclass
 class FileInfoFixture:
