+++ tests/sdk/python/rtdip_sdk/pipelines/destinations/spark/test_delta.py
@@ -18,7 +18,6 @@ import pytest
 from src.sdk.python.rtdip_sdk._sdk_utils.compare_versions import _get_package_version
 from src.sdk.python.rtdip_sdk.pipelines.destinations.spark.delta import SparkDeltaDestination
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, MavenLibrary
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql.functions import lit
 from pyspark.sql import SparkSession
 from pytest_mock import MockerFixture
