+++ tests/sdk/python/rtdip_sdk/pipelines/destinations/spark/test_pcdm_to_delta.py
@@ -22,7 +22,6 @@ from src.sdk.python.rtdip_sdk.pipelines.destinations.spark.delta import SparkDel
 from src.sdk.python.rtdip_sdk.pipelines.destinations.spark.pcdm_to_delta import SparkPCDMToDeltaDestination
 from src.sdk.python.rtdip_sdk.pipelines.utilities.spark.delta_table_create import DeltaTableCreateUtility, DeltaTableColumn
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, MavenLibrary
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql import SparkSession
 from pyspark.sql.functions import col
 from pytest_mock import MockerFixture
