+++ tests/sdk/python/rtdip_sdk/pipelines/sources/spark/iso/test_base_iso.py
@@ -18,7 +18,6 @@ from requests import HTTPError
 
 from src.sdk.python.rtdip_sdk.pipelines.sources.spark.iso import BaseISOSource
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql import DataFrame, SparkSession
 from pytest_mock import MockerFixture
