+++ tests/sdk/python/rtdip_sdk/pipelines/sources/spark/iso/test_miso_daily_load_iso.py
@@ -22,7 +22,6 @@ import pytest
 from src.sdk.python.rtdip_sdk.pipelines.sources.spark.iso import MISODailyLoadISOSource
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.iso import MISO_SCHEMA
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql import DataFrame, SparkSession
 from pytest_mock import MockerFixture
