+++ tests/sdk/python/rtdip_sdk/pipelines/sources/spark/test_iot_hub.py
@@ -18,7 +18,6 @@ import pytest
 from src.sdk.python.rtdip_sdk.pipelines.sources.spark.iot_hub import SparkIoThubSource
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, MavenLibrary
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark import EVENTHUB_SCHEMA
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 import json
 from pyspark.sql import DataFrame, SparkSession
 from pytest_mock import MockerFixture
