+++ tests/sdk/python/rtdip_sdk/pipelines/sources/spark/test_kafka.py
@@ -20,7 +20,6 @@ from src.sdk.python.rtdip_sdk._sdk_utils.compare_versions import _get_package_ve
 from src.sdk.python.rtdip_sdk.pipelines.sources.spark.kafka import SparkKafkaSource
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, MavenLibrary
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark import KAFKA_SCHEMA
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 import json
 from pyspark.sql import DataFrame, SparkSession
 from pytest_mock import MockerFixture
