+++ tests/sdk/python/rtdip_sdk/pipelines/sources/spark/weather/test_base_weather.py
@@ -17,7 +17,6 @@ from requests import HTTPError
 
 from src.sdk.python.rtdip_sdk.pipelines.sources.spark.weather import BaseWeatherSource
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql import DataFrame, SparkSession
 from pytest_mock import MockerFixture
