+++ tests/sdk/python/rtdip_sdk/pipelines/transformers/spark/iso/test_miso_to_mdm.py
@@ -22,7 +22,6 @@ from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.mdm import MDM_USAGE_SCH
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.iso import MISO_SCHEMA
 from src.sdk.python.rtdip_sdk.pipelines.transformers import MISOToMDMTransformer
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, SystemType
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql import SparkSession, DataFrame
