+++ tests/sdk/python/rtdip_sdk/pipelines/transformers/spark/test_pandas_to_pyspark.py
@@ -16,7 +16,6 @@ import sys
 sys.path.insert(0, '.')
 from src.sdk.python.rtdip_sdk.pipelines.transformers.spark.pandas_to_pyspark import PandasToPySparkTransformer
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, SystemType
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 
 from pyspark.sql import SparkSession
 from pyspark.sql.dataframe import DataFrame
