+++ tests/sdk/python/rtdip_sdk/pipelines/utilities/spark/test_delta_table_optimize.py
@@ -17,7 +17,6 @@ sys.path.insert(0, '.')
 
 from src.sdk.python.rtdip_sdk.pipelines.utilities.spark.delta_table_create import DeltaTableCreateUtility, DeltaTableColumn
 from src.sdk.python.rtdip_sdk.pipelines.utilities.spark.delta_table_optimize import DeltaTableOptimizeUtility
-from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 from pyspark.sql import SparkSession
 
 def test_spark_delta_table_optimize(spark_session: SparkSession):
